Separate Eigen device functions into individual files.
