+++ primitiv/device_ops/copy_tensor_eigen.cc
@@ -1,0 +1,23 @@
+#include <primitiv/config.h>
+
+#include <primitiv/eigen_device.h>
+#include <primitiv/device_ops/eigen_utils.h>
+
+namespace primitiv {
+namespace devices {
+
+void Eigen::copy_tensor_impl(const Tensor &x, Tensor &y) {
+  switch (x.device().type()) {
+    case Device::DeviceType::NAIVE:
+      reset_tensor_by_array(CDATA(x), y);
+      break;
+    case Device::DeviceType::EIGEN:
+      reset_tensor_by_array(CDATA(x), y);
+      break;
+    default:
+      reset_tensor_by_vector(x.to_vector(), y);
+  }
+}
+
+}  // namespace devices
+}  // namespace primitiv
