+++ primitiv/device_ops/elu_eigen.cc
@@ -1,0 +1,13 @@
+#include <primitiv/config.h>
+
+#include <primitiv/eigen_device.h>
+#include <primitiv/device_ops/eigen_utils.h>
+
+namespace primitiv {
+namespace devices {
+
+EIGEN_DEV_FW_X_CONST(elu, (x > 0.).select(x, k * (x.exp() - 1.)));
+EIGEN_DEV_BW_X_CONST(elu, (x > 0.).select(gy, (y + k) * gy));
+
+}  // namespace devices
+}  // namespace primitiv
