+++ primitiv/device_ops/identity_eigen.cc
@@ -1,0 +1,15 @@
+#include <primitiv/config.h>
+
+#include <primitiv/eigen_device.h>
+#include <primitiv/device_ops/eigen_utils.h>
+
+namespace primitiv {
+namespace devices {
+
+void Eigen::identity_impl(Tensor &y) {
+  const std::size_t size = y.shape()[0];
+  EMap<EMatrixXf>(MDATA(y), size, size).setIdentity();
+}
+
+}  // namespace devices
+}  // namespace primitiv
