+++ primitiv/device_ops/negate_eigen.cc
@@ -1,0 +1,12 @@
+#include <primitiv/config.h>
+
+#include <primitiv/eigen_device.h>
+#include <primitiv/device_ops/eigen_utils.h>
+
+namespace primitiv {
+namespace devices {
+
+EIGEN_DEV_FW_X(negate, -x);
+
+}  // namespace devices
+}  // namespace primitiv
