+++ primitiv/device_ops/random_normal_eigen.cc
@@ -1,0 +1,14 @@
+#include <primitiv/config.h>
+
+#include <primitiv/eigen_device.h>
+#include <primitiv/device_ops/eigen_utils.h>
+
+namespace primitiv {
+namespace devices {
+
+void Eigen::random_normal_impl(float mean, float sd, Tensor &y) {
+  randomizer_.fill_normal(mean, sd, y.shape().size(), MDATA(y));
+}
+
+}  // namespace devices
+}  // namespace primitiv
