+++ primitiv/device_ops/random_uniform_eigen.cc
@@ -1,0 +1,14 @@
+#include <primitiv/config.h>
+
+#include <primitiv/eigen_device.h>
+#include <primitiv/device_ops/eigen_utils.h>
+
+namespace primitiv {
+namespace devices {
+
+void Eigen::random_uniform_impl(float lower, float upper, Tensor &y) {
+  randomizer_.fill_uniform(lower, upper, y.shape().size(), MDATA(y));
+}
+
+}  // namespace devices
+}  // namespace primitiv
