+++ primitiv/device_ops/reset_tensor_eigen.cc
@@ -1,0 +1,14 @@
+#include <primitiv/config.h>
+
+#include <primitiv/eigen_device.h>
+#include <primitiv/device_ops/eigen_utils.h>
+
+namespace primitiv {
+namespace devices {
+
+void Eigen::reset_tensor_impl(float k, Tensor &x) {
+  EMap<EArrayXf>(MDATA(x), x.shape().size()).setConstant(k);
+}
+
+}  // namespace devices
+}  // namespace primitiv
