+++ primitiv/device_ops/sum_eigen.cc
@@ -1,14 +1,14 @@
 #include <primitiv/config.h>
 
-#include <cmath>
-
-#include <primitiv/naive_device.h>
-#include <primitiv/device_ops/naive_utils.h>
+#include <primitiv/eigen_device.h>
+#include <primitiv/device_ops/eigen_utils.h>
 
 namespace primitiv {
 namespace devices {
 
-void Naive::logsumexp_fw_impl(const Tensor &x, std::uint32_t dim, Tensor &y) {
+void Eigen::sum_fw_impl(const Tensor &x, std::uint32_t dim, Tensor &y) {
+  // TODO(odashi): Optimize this functions using Eigen operations.
+
   const std::uint32_t n = x.shape()[dim];
   const std::uint32_t repeat = y.shape().size();
   const std::uint32_t skip1 = y.shape().lower_volume(dim);
@@ -16,15 +16,11 @@ void Naive::logsumexp_fw_impl(const Tensor &x, std::uint32_t dim, Tensor &y) {
   float *dest = MDATA(y);
   const float *src = CDATA(x);
   for (std::uint32_t i = 0; i < repeat; ++i) {
-    // TODO(odashi): This calculation might generate large errors.
     std::uint32_t offset = i % skip1 + (i / skip1) * skip2;
-    float tmp = src[offset];
-    for (std::uint32_t j = 1; j < n; ++j) {
+    float tmp = 0;
+    for (std::uint32_t j = 0; j < n; ++j) {
+      tmp += src[offset];
       offset += skip1;
-      float arg = src[offset];
-      tmp = tmp > arg
-        ? tmp + std::log(1. + std::exp(arg - tmp))
-        : arg + std::log(1. + std::exp(tmp - arg));
     }
     dest[i] = tmp;
   }
