+++ primitiv/device_ops/tan_eigen.cc
@@ -1,0 +1,13 @@
+#include <primitiv/config.h>
+
+#include <primitiv/eigen_device.h>
+#include <primitiv/device_ops/eigen_utils.h>
+
+namespace primitiv {
+namespace devices {
+
+EIGEN_DEV_FW_X(tan, x.tan());
+EIGEN_DEV_BW_X(tan, gy * (1. + y * y));
+
+}  // namespace devices
+}  // namespace primitiv
